backend/pyproject.toml
[project]
name = "antiplagism"
version = "0.1.0"
description = "Diploma project for TSUC IET"
authors = [
    {name = "Fakhriddin",email = "fahriddin3040@gmail.com"}
]
license = {text = "Apache"}
readme = "README.md"
requires-python = ">=3.10, <=3.13.2"
dependencies = [
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "fastapi (>=0.115.10,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
black = "^25.1.0"
ruff = "^0.9.9"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
